+++ packages/dharma-core/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./lib/createStore";
 export * from "./lib/merge";
+export * from "./lib/types";
