+++ packages/dharma-react/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./lib/createStoreContext";
+export * from "./lib/types";
 export * from "./lib/useStore";
 export * from "./lib/useStoreContext";
